backend/requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
mysqlclient==2.2.4
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
